fix: widen lib-client amount parameters from u64 to u128
All token/transfer/stake builder functions now accept u128 for amount
parameters. This is required for 18-decimal SOV where amounts like
5000 SOV = 5×10^21 overflow u64.

Changed:
- lib-client/src/token_tx.rs: build_transfer_tx, build_sov_wallet_transfer_tx,
  build_token_wallet_transfer_tx, build_mint_tx, build_create_token_tx,
  build_burn_tx — all amount params u64 → u128
- lib-client/src/dao_tx.rs: build_treasury_allocation_tx, build_dao_stake_tx,
  build_dao_unstake_tx — amount params u64 → u128
- lib-client/src/bonding_curve_tx.rs: build_add_liquidity_tx — amounts u64 → u128
- lib-client/src/lib.rs: all FFI wrapper functions widened
- zhtp-cli/src/commands/domain.rs: fee_amount_atoms u64 → u128

Note: TreasuryAllocationData.amount (blockchain struct) remains u64 with
safe downcast at boundary. DaoStakeData.amount is u128 (already widened).

+++ lib-client/src/bonding_curve_tx.rs
@@ -382,8 +382,8 @@ pub fn build_add_liquidity_tx(
     identity: &crate::Identity,
     token_id: &[u8; 32],
     pool_id: &[u8; 32],
-    token_amount: u64,
-    sov_amount: u64,
+    token_amount: u128,
+    sov_amount: u128,
     chain_id: u8,
     nonce: u64,
 ) -> Result<String, String> {
