+++ lib-client/src/lib.rs
@@ -1044,7 +1044,7 @@ pub extern "C" fn zhtp_client_build_token_transfer(
     token_id: *const u8,
     to_pubkey: *const u8,
     to_pubkey_len: usize,
-    amount: u64,
+    amount: u128,
     chain_id: u8,
     nonce: u64,
 ) -> *mut std::ffi::c_char {
@@ -1091,7 +1091,7 @@ pub extern "C" fn zhtp_client_build_sov_wallet_transfer(
     handle: *const IdentityHandle,
     from_wallet_id: *const u8,
     to_wallet_id: *const u8,
-    amount: u64,
+    amount: u128,
     chain_id: u8,
     nonce: u64,
 ) -> *mut std::ffi::c_char {
@@ -1138,7 +1138,7 @@ pub extern "C" fn zhtp_client_build_token_wallet_transfer(
     token_id: *const u8,
     from_wallet_id: *const u8,
     to_wallet_id: *const u8,
-    amount: u64,
+    amount: u128,
     chain_id: u8,
     nonce: u64,
 ) -> *mut std::ffi::c_char {
@@ -1417,7 +1417,7 @@ pub extern "C" fn zhtp_client_build_token_mint(
     token_id: *const u8,
     to_pubkey: *const u8,
     to_pubkey_len: usize,
-    amount: u64,
+    amount: u128,
     chain_id: u8,
 ) -> *mut std::ffi::c_char {
     if handle.is_null() || token_id.is_null() || to_pubkey.is_null() {
@@ -1507,7 +1507,7 @@ pub extern "C" fn zhtp_client_build_token_create(
 pub extern "C" fn zhtp_client_build_token_burn(
     handle: *const IdentityHandle,
     token_id: *const u8,
-    amount: u64,
+    amount: u128,
     chain_id: u8,
 ) -> *mut std::ffi::c_char {
     if handle.is_null() || token_id.is_null() {
@@ -1557,7 +1557,7 @@ pub extern "C" fn zhtp_client_build_token_burn(
 pub unsafe extern "C" fn zhtp_client_build_dao_stake(
     handle: *const IdentityHandle,
     sector_dao_key_id: *const u8,
-    amount: u64,
+    amount: u128,
     nonce: u64,
     lock_blocks: u64,
     chain_id: u8,
@@ -1863,7 +1863,7 @@ pub extern "C" fn zhtp_client_build_bonding_curve_deploy(
 pub extern "C" fn zhtp_client_build_bonding_curve_buy(
     handle: *const IdentityHandle,
     token_id: *const u8,
-    stable_amount: u64,
+    stable_amount: u128,
     min_tokens_out: u64,
     chain_id: u8,
     nonce: u64,
@@ -1908,7 +1908,7 @@ pub extern "C" fn zhtp_client_build_bonding_curve_buy(
 pub extern "C" fn zhtp_client_build_bonding_curve_sell(
     handle: *const IdentityHandle,
     token_id: *const u8,
-    token_amount: u64,
+    token_amount: u128,
     min_stable_out: u64,
     chain_id: u8,
     nonce: u64,
@@ -2009,8 +2009,8 @@ pub extern "C" fn zhtp_client_build_add_liquidity(
     handle: *const IdentityHandle,
     token_id: *const u8,
     pool_id: *const u8,
-    token_amount: u64,
-    sov_amount: u64,
+    token_amount: u128,
+    sov_amount: u128,
     chain_id: u8,
     nonce: u64,
 ) -> *mut std::ffi::c_char {
